# server/shopper.openapi.yml
openapi: 3.0.1
info:
  title: Shopper API.
  description: Complete documentation for the Shopper API.
  version: 1.0.0
servers:
  - url: http://api.shopper.ink/v1
paths:
...
  /bcds: # Add / Edit / Delete article
    post:
      deprecated: true
      summary: Add article to the barcode database
      tags:
        - Barcode Data Store
      requestBody:
        required: true
        content:
            application/json:
              schema:
                type: object
                properties:
                  barcode:
                    type: string
                    example: 7310070001764
                  display_name:
                    type: string
                    example: Ramlösa Smultron 50cl
                  manufacturer:
                    type: string
                    example: Carlsberg Sweden
                required:
                  - barcode
                  - display_name
                  - manufacturer
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /bcds/{barcode}:
    get:
      summary: Get info about a specific article
      tags:
        - Barcode Data Store
      responses:
        '200':
          description: Article object
          content:
            application/json:
              schema:
                type: object
                properties:
                  barcode:
                    type: string
                    example: 7310070001764
                  display_name:
                    type: string
                    example: Ramlösa Smultron 50cl
                  manufacturer:
                    type: string
                    example: Carlsberg Sweden
                required:
                  - barcode
                  - display_name
                  - manufacturer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      parameters:
        - name: barcode
          in: path
          description: Barcode string
          required: true
          schema:
            type: string
    patch:
      deprecated: true
      summary: Edit an article in the barcode database
      tags:
        - Barcode Data Store
      requestBody:
        required: true
        content:
            application/json:
              schema:
                type: object
                properties:
                  barcode:
                    type: string
                    example: 7310070001764
                  display_name:
                    type: string
                    example: Ramlösa Smultron 50cl
                  manufacturer:
                    type: string
                    example: Carlsberg Sweden
                required:
                  - barcode
                  - display_name
                  - manufacturer
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters:
        - name: barcode
          in: path
          description: Barcode string
          required: true
          schema:
            type: string
    delete:
      deprecated: true
      summary: Delete article from the barcode database
      tags:
        - Barcode Data Store
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters:
        - name: barcode
          in: path
          description: Barcode string
          required: true
          schema:
            type: string
components:
  responses:
    Ok: # 200
      description: Ok
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericResponse'
    Created: # 201
      description: Resource created
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericResponse'
    BadRequest: # 400
      description: Malformed request syntax
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericResponse'
    Unauthorized: # 401
      description: Valid request but credentials are incorrect
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericResponse'
    NotFound: # 404
      description: Resourse not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericResponse'
    Conflict: # 409
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericResponse'
...
  schemas:
...
    GenericResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - Bearer: []